Add prev page btn on cont. token pages through JS
This commit adds the ability to fully paginate pages
that rely on continuation tokens. This works by caching the tokens
within session storage and fetching it via JS.

The caveat of course is that non-js users wouldn't be able to
utilize the previous page buttons and instead will have the behavior
currently seen on Invidious.